Fixed/Added Some stuff
Fixed some of the maps so that they associate correctly.
Fixed the game session folder creation to not create new folders if we are still in the game. It does this by checking if we had any match or round info inside the current file we are reading and if so we can assume its the same game session, so reuse the previously made gamesession file, instead of creating a new one
Added so that it also cleans up old game session folders which were created and left empty with no round or match info.
Fixed a bug that if you left prematurely in the match when you know you wont qualify for the next round, it doenst delete that info now, and maintains it, knowing that you basically are leaving cause you have no chance of moving forward to the next round.
Tested alot of scenarios for the game session and made sure that it handles all types of random user behaviours.
Made it so that if you stop the program mid game session it can continue to run the program appropriately
Fixed that if the game crashes it reconigzes this and deletes the old info from the crash as its now qualified as a premature exit.
And alot more smaller things.
Added more to do, and and cleaned it up a bit