Fixed alot of bugs
Fixed alot of weird issues and bugs that kept happening, as of right now it seems to be working perfectly, I have tried to break it in many ways but it seems to be fine. Last thing to add is the duration and the program then is feature complete and ready.